+++ CHANGELOG.md
@@ -7,7 +7,9 @@
 
 ### New features:
 * Add modules for first and second order Fermi acceleration
+* Make custom photon fields available in the PhotoPionProduction module.
 * Add tricubic- and nearest neighbour interpolation routines for scalar- and vectorgrids.
+* Add the new PolarizedSingleModeMagneticField class for polarized/helical single mode magnetic field models.
 
 ### Interface changes: